Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int lli;
 	float f;
 
-	printf("Size of a char: %s bytes(s)\n", sizeof(c));
-	printf("Size of an int: %i bytes(s)\n", sizeof(i));
-	printf("Size of a long int: %li byte(s)\n", sizeof(li));
-	printf("Size of a long long int: %lli byte(s)\n", sizeof(lli));
-	printf("Size of a float: %f bytes(s)\n", sizeof(f));
+	printf("Size of a char: %s bytes(s)\n",(unsigned long)sizeof(c));
+	printf("Size of an int: %i bytes(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %li byte(s)\n", (unsigned long)sizeof(li));
+	printf("Size of a long long int: %lli byte(s)\n", (unsigned long)sizeof(lli));
+	printf("Size of a float: %f bytes(s)\n", (unsigned long)sizeof(f));
 
 	return (0);
 }
